Move Elasticsearch traits
