+++ app/Models/Beatmapset.php
@@ -103,7 +103,7 @@ use Illuminate\Database\QueryException;
  */
 class Beatmapset extends Model implements AfterCommit, Commentable, Indexable
 {
-    use CommentableDefaults, Elasticsearch\BeatmapsetTrait, Memoizes, Traits\Reportable, SoftDeletes, Validatable;
+    use CommentableDefaults, Traits\Es\BeatmapsetSearch, Memoizes, Traits\Reportable, SoftDeletes, Validatable;
 
     protected $_storage = null;
     protected $table = 'osu_beatmapsets';
