+++ app/Models/Traits/Es/BaseIndexable.php
@@ -3,12 +3,12 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Traits;
+namespace App\Models\Traits\Es;
 
 use App\Libraries\Elasticsearch\Es;
 use DateTime;
 
-trait EsIndexable
+trait BaseIndexable
 {
     abstract public static function esIndexName();
