+++ app/Models/Traits/Es/ForumPostSearch.php
@@ -3,15 +3,14 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Models\Elasticsearch;
+namespace App\Models\Traits\Es;
 
 use App\Models\Forum\Forum;
-use App\Traits\EsIndexableModel;
 use Carbon\Carbon;
 
-trait PostTrait
+trait ForumPostSearch
 {
-    use EsIndexableModel;
+    use BaseDbIndexable;
 
     public static function esIndexName()
     {
