+++ app/Models/Traits/Es/UserSearch.php
@@ -3,14 +3,13 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Models\Elasticsearch;
+namespace App\Models\Traits\Es;
 
-use App\Traits\EsIndexableModel;
 use Carbon\Carbon;
 
-trait UserTrait
+trait UserSearch
 {
-    use EsIndexableModel;
+    use BaseDbIndexable;
 
     public static function esIndexName()
     {
