+++ app/Models/Traits/Es/WikiPageSearch.php
@@ -3,13 +3,11 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Models\Elasticsearch;
+namespace App\Models\Traits\Es;
 
-use App\Traits\EsIndexable;
-
-trait WikiPageTrait
+trait WikiPageSearch
 {
-    use EsIndexable;
+    use BaseIndexable;
 
     public static function esIndexName()
     {
